fix(server): renaming main to server
